requirements.txt
langchain==0.3.18
langchain-google-genai==2.0.2
google-generativeai==0.8.3
boto3
python-dotenv
fastapi
uvicorn
pydantic>=2.0.0,<3.0.0
faiss-cpu
unstructured